+++ AWS/cloudfront/main.tf
@@ -9,7 +9,7 @@ locals {
 
 module "cloudfront" {
   source  = "terraform-aws-modules/cloudfront/aws"
-  version = "~> 4.0"
+  version = "~> 6.4.1"
 
   for_each = var.distributions